requirements.txt
django==4.2.5
gunicorn==19.9.0
django_filter==23.2
django-bootstrap4==23.2
requests==2.31.0
djangorestframework==3.14.0
django-rest-swagger==2.2.0
drf-spectacular==0.27.1
celery[redis]==5.3.4
redis==4.5.2
django-redis==4.12.1

ipython
